Refactor Upower
